% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blast_tab.R
\name{read_blast_tab}
\alias{read_blast_tab}
\title{Read BLAST tabular output}
\usage{
read_blast_tab(path)
}
\arguments{
\item{path}{string. Path to BLAST tab file.}
}
\value{
tibble.
}
\description{
Reads the BLAST tabular format generated by -outfmt 6 into an R tibble with
named columns.
}
\details{
Column names correspond to:
qseqid: query (e.g., gene) sequence id
sseqid: subject (e.g., reference genome) sequence id
pident: percentage of identical matches
length: alignment length
mismatch: number of mismatches
gapopen: number of gap openings
qstart: start of alignment in query
qend: end of alignment in query
sstart: start of alignment in subject
send: end of alignment in subject
evalue: expect value
bitscore: bit score
}
